Added a couple of badges.
Also added the content documents to update.

Chris will update the top of the web page, and get me the maritime FAQ.

I will redo the "hero unit" to be quite a bit smaller, probably 2 columns with the title in real text with space for a commercial video he'll post to Vimeo. (Maybe 2 of them...)

The two column details for maritime and personal injury sections will use a disclosure mechanism. Probably a hand selected "summary" that will expand down via a click, probably a wee bit of jS.

I'll also fix up the alignment of the badges.